requirements.txt
streamlit
gspread
google-auth
pandas
plotly
fpdf
